Create JavaOutputFormatting-PracticeWithPrintf
The first column should be left justified using exactly 15 characters. The integer of the second column should have exactly 3 digits. If the original input has less than 3 digits, you should pad with zeros to the left.

To make the problem easier, some part of the solution is already given in the editor, just complete the remaining parts.